auth_module/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.30
alembic>=1.13.2
psycopg2-binary>=2.9.9
pydantic>=2.7.0
jwcrypto>=1.5.6
pyjwt[crypto]>=2.9.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
starlette>=0.40.0
httpx>=0.27.0